// src/domain/value-objects/NodeType.d.ts
/**
 * NodeType - Types of nodes in the property graph
 */
export declare enum NodeType {
    FILE = "file",
    DIRECTORY = "directory",
    MODULE = "module",
    CLASS = "class",
    INTERFACE = "interface",
    FUNCTION = "function",
    METHOD = "method",
    VARIABLE = "variable",
    CONSTANT = "constant",
    ENUM = "enum",
    TYPE = "type",
    IMPORT = "import",
    EXPORT = "export"
}
//# sourceMappingURL=NodeType.d.ts.map