// src/domain/value-objects/EdgeType.d.ts
/**
 * EdgeType - Types of edges in the property graph
 */
export declare enum EdgeType {
    DEPENDS_ON = "depends_on",
    IMPORTS = "imports",
    EXPORTS = "exports",
    CONTAINS = "contains",
    CALLS = "calls",
    EXTENDS = "extends",
    IMPLEMENTS = "implements",
    REFERENCES = "references"
}
//# sourceMappingURL=EdgeType.d.ts.map